aceso-app/aceso-api/requirements.txt
absl-py==0.7.1
adal==1.2.1
ansiwrap==0.8.4
applicationinsights==0.11.9
asn1crypto==0.24.0
astor==0.8.0
attrs==19.1.0
Automat==0.6.0
azure-common==1.1.22
azure-graphrbac==0.61.1
azure-mgmt-authorization==0.52.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-keyvault==2.0.0
azure-mgmt-resource==3.0.0
azure-mgmt-storage==4.0.0
azureml-contrib-notebook==1.0.43
azureml-core==1.0.43.1
azureml-dataprep==1.1.5
azureml-dataprep-native==13.0.0
azureml-pipeline==1.0.43
azureml-pipeline-core==1.0.43
azureml-pipeline-steps==1.0.43
azureml-sdk==1.0.43
azureml-telemetry==1.0.43.1
azureml-train==1.0.43
azureml-train-core==1.0.43
azureml-train-restclients-hyperdrive==1.0.43
azureml-widgets==1.0.43.1
backcall==0.1.0
backports.tempfile==1.0
backports.weakref==1.0.post1
basemap==1.2.0
beautifulsoup4==4.6.0
bleach==3.1.0
blinker==1.4
cachetools==3.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloud-init==18.2
cloudpickle==0.5.3
cmake==3.12.0
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.7
cycler==0.10.0
dask==0.18.2
decorator==4.4.0
defusedxml==0.6.0
distro==1.4.0
distro-info==0.18
dlib==19.15.0
docker==4.0.2
dotnetcore2==2.1.8
entrypoints==0.3
Flask==1.0.2
Flask-Cors==3.0.7
future==0.17.1
gast==0.2.2
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-vision==0.35.2
google-pasta==0.1.7
googleapis-common-protos==1.5.6
googlemaps==3.0.2
grpcio==1.21.1
h5py==2.9.0
html5lib==1.0.1
httplib2==0.9.2
hyperlink==17.3.1
idna==2.8
imageio==2.5.0
imutils==0.5.1
incremental==16.10.1
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.14.0
jeepney==0.4
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
jsonpatch==1.16
jsonpickle==1.2
jsonpointer==1.10
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==5.2.0
jupyter-core==4.5.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
kmeans==1.0.2
language-selector==0.1
lesscpy==0.13.0
lib==3.0.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mistune==0.8.4
mpg123==0.4
msrest==0.6.7
msrestazure==0.6.1
nbconvert==5.5.0
nbformat==4.4.0
ndg-httpsclient==0.5.1
networkx==2.3
notebook==5.7.8
numpy==1.16.4
oauthlib==3.0.1
olefile==0.45.1
onnx==1.5.0
onnxconverter-common==1.5.0
opencv-python==3.4.2.17
PAM==0.4.2
pandas==0.23.3
pandocfilters==1.4.2
papermill==1.0.1
parso==0.5.0
pathspec==0.5.9
pexpect==4.7.0
php==1.2.0
pickleshare==0.7.5
Pillow==6.0.0
playsound==1.2.2
plib3.stdlib==1.0.3
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.8.0
ptyprocess==0.6.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
pycrypto==2.6.1
Pygments==2.4.2
pygobject==3.26.1
PyJWT==1.7.1
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyproj==1.9.5.1
pyrsistent==0.15.2
pyserial==3.4
pyshp==2.1.0
PySocks==1.6.8
python-apt==1.6.0
python-dateutil==2.8.0
python-debian==0.1.32
pytz==2019.1
PyWavelets==1.0.3
pyxdg==0.25
PyYAML==5.1.1
pyzmq==18.0.1
qtconsole==4.4.1
requests==2.22.0
requests-oauthlib==1.2.0
requests-unixsocket==0.1.5
rsa==4.0
ruamel.yaml==0.15.89
scikit-image==0.15.0
scikit-learn==0.21.2
scikit-video==1.1.11
scipy==1.3.0
SecretStorage==3.1.1
selenium==3.13.0
Send2Trash==1.5.0
service-identity==16.0.0
simplegeneric==0.8.1
six==1.12.0
skl2onnx==1.5.0
splinter==0.8.0
spotipy==2.4.4
ssh-import-id==5.7
systemd-python==234
tenacity==5.0.4
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
textwrap3==0.9.2
toolz==0.9.0
tornado==6.0.2
tqdm==4.32.2
traitlets==4.3.2
twilio==6.17.0
Twisted==17.9.0
typing==3.7.4
typing-extensions==3.7.4
ufw==0.35
unattended-upgrades==0.1
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.15.4
widgetsnbextension==3.4.2
wrapt==1.11.2
zope.interface==4.3.2
